Add files via upload
This case study seeks to illustrate the practical application of Exploratory Data Analysis (EDA) within a real-world business context. In addition to employing established EDA techniques, the study will delve into the realm of risk analytics in the banking and financial services sector. The primary objective is to comprehend how data is harnessed to mitigate the risk of financial losses when extending loans to customers. Through this analysis, we aim to gain valuable insights into the factors influencing lending risk and contribute to informed decision-making processes in the financial industry.